Add docker-compose for manager debug
Also adds insert into Contact table for db initialization, and adds a
throw to NHibernateExtensions to clearly show when the connection string
is empty.

Next step: figure out how to better attach the debugger to the docker
container and figure out why no Watch Items are being picked up in the
WatchItemManager.